Move page files to page folder
